# docs/openapi_v1.yaml
openapi: 3.0.3
info:
  title: UtilBase Mobile API
  version: "1.0.0"
  description: Minimal contract for Android executor app.
servers:
  - url: /api/v1
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login and get access/refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Tokens issued
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and issue a new access token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tokens rotated
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token revoked
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
...
  /requests/{id}/checklist-template:
    get:
      tags: [checklist]
      summary: Get resolved checklist template for request
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/requestId"
      responses:
        "200":
          description: Checklist template and answers
  /requests/{id}/checklist-submit:
    post:
      tags: [checklist]
      summary: Submit checklist answers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/requestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id:
                        type: integer
                      checked:
                        type: boolean
                      value_text:
                        type: string
                      value_number:
                        type: number
                      media_id:
                        type: integer
      responses:
        "200":
          description: Checklist state updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    requestId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...